pyproject.toml
[project]
name = "devspace-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.72.0",
    "gradio>=6.11.0",
]
